+++ pkger/parser_test.go
@@ -121,9 +121,10 @@ spec:
 
 				actual := buckets[0]
 				expectedBucket := SummaryBucket{
-					Name:            "rucket_11",
-					Description:     "bucket 1 description",
-					RetentionPeriod: time.Hour,
+					Name:              "rucket_11",
+					Description:       "bucket 1 description",
+					RetentionPeriod:   time.Hour,
+					LabelAssociations: []SummaryLabel{},
 				}
 				assert.Equal(t, expectedBucket, actual)
 			})